+++ src/main/java/coffeemeet/server/admin/domain/AdminRepository.java
@@ -1,6 +1,5 @@
-package coffeemeet.server.admin.infrastructure;
+package coffeemeet.server.admin.domain;
 
-import coffeemeet.server.admin.domain.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdminRepository extends JpaRepository<Admin, String> {
